requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
regex==2024.11.6
tinydb==4.8.2
typing_extensions==4.12.2
Werkzeug==3.1.3
